+++ src/main/java/uk/gov/hmcts/reform/dg/docassembly/rest/TemplateRenditionResource.java
@@ -23,7 +23,7 @@ import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.dg.docassembly.config.Constants;
 import uk.gov.hmcts.reform.dg.docassembly.dto.CreateTemplateRenditionDto;
 import uk.gov.hmcts.reform.dg.docassembly.service.TemplateRenditionService;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.DocumentTaskProcessingException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentTaskProcessingException;
 
 import java.io.IOException;
