+++ src/main/java/uk/gov/hmcts/reform/dg/docassembly/service/DmStoreDownloader.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.dg.docassembly.service;
 
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.DocumentTaskProcessingException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentTaskProcessingException;
 
 import java.io.File;
