+++ src/main/java/uk/gov/hmcts/reform/dg/docassembly/service/impl/FileToPDFConverterServiceImpl.java
@@ -9,9 +9,9 @@ import uk.gov.hmcts.reform.dg.docassembly.conversion.DocmosisConverter;
 import uk.gov.hmcts.reform.dg.docassembly.service.CdamService;
 import uk.gov.hmcts.reform.dg.docassembly.service.DmStoreDownloader;
 import uk.gov.hmcts.reform.dg.docassembly.service.FileToPDFConverterService;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.DocumentProcessingException;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.DocumentTaskProcessingException;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.FileTypeException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentProcessingException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentTaskProcessingException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.FileTypeException;
 
 import java.io.File;
 import java.io.IOException;
