+++ src/test/java/uk/gov/hmcts/reform/dg/docassembly/rest/DocumentConversionResourceTest.java
@@ -9,8 +9,8 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.http.ResponseEntity;
 import uk.gov.hmcts.reform.dg.docassembly.service.FileToPDFConverterService;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.DocumentProcessingException;
-import uk.gov.hmcts.reform.dg.docassembly.service.exception.FileTypeException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentProcessingException;
+import uk.gov.hmcts.reform.dg.docassembly.exception.FileTypeException;
 
 import java.io.File;
 import java.util.UUID;
