+++ src/test/java/uk/gov/hmcts/reform/dg/docassembly/service/DmStoreUploaderTest.java
@@ -12,6 +12,7 @@ import uk.gov.hmcts.reform.auth.checker.core.user.UserResolver;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.dg.docassembly.dto.CreateTemplateRenditionDto;
 import uk.gov.hmcts.reform.dg.docassembly.dto.RenditionOutputType;
+import uk.gov.hmcts.reform.dg.docassembly.exception.DocumentUploaderException;
 
 import java.io.File;
 import java.util.Base64;
